Split type checker calls.rs into builtin_calls and static_dispatch modules
